Break the loop when failed is set to true. Also, use a stack pointer if the passed failed pointer is NULL.

+++ xpdev/xpprintf.c
@@ -3,7 +3,7 @@
 
 /* Deuce's vs[n]printf() replacement */
 
-/* $Id: xpprintf.c,v 1.61 2018/03/10 06:42:41 deuce Exp $ */
+/* $Id: xpprintf.c,v 1.62 2018/04/07 07:21:46 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -725,7 +725,7 @@ char* DLLCALL xp_asprintf_next(char *format, int type, ...)
 							i=(int)ld;
 							break;
 						case XP_PRINTF_TYPE_VOIDP:
-							i=(int)pntr;
+							i=(long)pntr;
 							break;
 						case XP_PRINTF_TYPE_SIZET:
 							i=s;
@@ -764,7 +764,7 @@ char* DLLCALL xp_asprintf_next(char *format, int type, ...)
 							i=(int)ld;
 							break;
 						case XP_PRINTF_TYPE_VOIDP:
-							i=(int)pntr;
+							i=(long)pntr;
 							break;
 						case XP_PRINTF_TYPE_SIZET:
 							i=s;
@@ -803,7 +803,7 @@ char* DLLCALL xp_asprintf_next(char *format, int type, ...)
 							ui=(unsigned)ld;
 							break;
 						case XP_PRINTF_TYPE_VOIDP:
-							ui=(unsigned int)pntr;
+							ui=(unsigned long)pntr;
 							break;
 						case XP_PRINTF_TYPE_SIZET:
 							ui=s;
